+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added `tag` prop to the `Results` component
 - Simplified the `Checkbox` and `Radio` components. Removed all the extra markup and will now simply render an input field
 - Removed the `options` prop from the `Select` component. Can now be used the same ways as the native select.
+- Removed the `results-path` prop from the `Search` component. Use the `path` prop of the `Results` and `NoResults` components.
 
 ## 1.1.0
 - The fields used for the label and value of Checkbox/Radio/Select components